extracted extenal_service from external_storage_service
